+++ app/src/main/kotlin/space/zuoqiu/siyuanhelper/MainActivity.kt
@@ -6,10 +6,10 @@ import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import space.zuoqiu.siyuanhelper.components.Editor
 import space.zuoqiu.siyuanhelper.ui.theme.SiYuanHelperTheme
 
 class MainActivity : ComponentActivity() {
@@ -22,25 +22,17 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    Greeting("Android")
+                    Editor("**粗体** *斜体* `行内代码` ~~删除线~~")
                 }
             }
         }
     }
 }
 
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
-
 @Preview(showBackground = true)
 @Composable
 fun GreetingPreview() {
     SiYuanHelperTheme {
-        Greeting("Android")
+        Editor("**粗体** *斜体* `行内代码` ~~删除线~~")
     }
 }