+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 更改日志 | CHANGE LOG
 
 - 项目初始化 | Project initialization
+- 添加 Markdown 编辑组件 | Add Markdown editor component